usb: gadget: f_fs: Add ioctl for allocating endpoint buffers.
This creates an ioctl named FUNCTIONFS_ENDPOINT_ALLOC which will
preallocate buffers for a given size. Any synchronous reads/writes
on that endpoint below that size will use those buffers instead of
allocating their own. If the endpoint is not active, the buffer
will not be allocated until it becomes active.

Signed-off-by: Jerry Zhang <zhangjerry@google.com>
Change-Id: Iddc2eb9e33bd4c819fa4201b2a43ffdcd54ee415