msm: qpnp-haptic: Reduce sysfs output of vtg_level to the raw mv value
When reading the vtg_level it is expected to return a string
value that directly represents the currently set mv value.
Kernel changes by asus modified the output format, leading to
an unexpected value.

Strip the output down to restore default behavior so LineageHW
and kernel managers can properly parse it.

Signed-off-by: Planet-X <planetx.dev@gmail.com>
Change-Id: I4e89547b22fd56be05171386355753d83a6bcc31