zenfone3: Remove asusdebug from the kernel
Asusdebug is mainly used to collect additional logs that are then saved to
the asdf partition. These logs are collected by asus services for sending
bug reports. As these asus services are not present on custom roms and
most of these messages are printed to dmesg anyways asusdebug can be removed.

This commit removes asusdebug. Useful debug messages were adapted so that
they are printed to dmesg instead of /asdf/ASUSEvtlog.txt.

Signed-off-by: Planet-X <planetx.dev@gmail.com>
Change-Id: Ie6546bfd5d1f60d362cc835060fae9f8ff464a6c